backend/requirements.txt
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.24.0
TTS>=0.21.0
yt-dlp>=2023.12.30